# docs/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete (deactivate) a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/roles": {
      "put": {
        "operationId": "UserController_setUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the user's assigned roles",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/cultivation/{locationId}/rooms": {
      "post": {
        "operationId": "RoomController_createRoom",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "description": "UUID of the location (license) where the room will be created",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing room name"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient role or module permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new cultivation room for a location",
        "tags": [
          "Cultivation � Rooms"
        ]
      },
      "get": {
        "operationId": "RoomController_listRooms",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "description": "UUID of the location to list rooms for",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of rooms returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient role or module permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cultivation rooms for a location",
        "tags": [
          "Cultivation � Rooms"
        ]
      }
    },
    "/cultivation/{ubi}/plants/{plantId}/cure": {
      "post": {
        "operationId": "CureController_curePlant",
        "parameters": [
          {
            "name": "ubi",
            "required": true,
            "in": "path",
            "description": "Location UBI (parent location identifier)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plantId",
            "required": true,
            "in": "path",
            "description": "UUID of the plant to cure",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cure record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or constraint violation"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Create a cure record for a plant (move from harvest -> cure workflow)",
        "tags": [
          "Cultivation � Cure"
        ]
      }
    }
  },
  "info": {
    "title": "Seed to Sale API",
    "description": "API documentation for Seed to Sale system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User UUID",
            "example": "11111111-1111-1111-1111-111111111111"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "name": {
            "type": "object",
            "description": "User full name",
            "example": "John Doe"
          },
          "roles": {
            "description": "List of role names assigned to the user",
            "example": [
              "licensee_admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locationIds": {
            "description": "List of location UUIDs the user is associated with",
            "example": [
              "33333333-3333-3333-3333-333333333333"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Account status (active/inactive)",
            "example": "active"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp when the user was created"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp when the user was last updated"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "New password. Provide only to change password via update (admin flow). Use change-password endpoint for user-initiated changes.",
            "minLength": 8,
            "writeOnly": true
          },
          "roleIds": {
            "description": "Array of role IDs to replace the user roles. To clear all roles pass an empty array.",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "locationIds": {
            "description": "Array of location IDs to replace the user locations. To clear pass an empty array.",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "parentLocationId": {
            "type": "string",
            "description": "Optional parent location id (UUID)",
            "example": "44444444-4444-4444-4444-444444444444"
          },
          "status": {
            "type": "string",
            "description": "Optional account status",
            "example": "active"
          },
          "isActive": {
            "type": "boolean",
            "description": "Optional toggle to activate/deactivate account",
            "example": true
          }
        }
      },
      "RoleAssignmentDto": {
        "type": "object",
        "properties": {
          "roleIds": {
            "description": "Array of role UUIDs to set for the user (replaces existing roles).",
            "example": [
              "11111111-1111-1111-1111-111111111111"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "roleIds"
        ]
      },
...
      "CreateCureDto": {
        "type": "object",
        "properties": {
          "dryFlowerWeight": {
            "type": "number",
            "description": "Dry weight of flower in grams",
            "example": 120.5,
            "minimum": 0
          },
          "dryOtherMaterialWeight": {
            "type": "number",
            "description": "Dry weight of other material (trim, stems, etc.) in grams",
            "example": 10,
            "minimum": 0
          },
          "dryWasteWeight": {
            "type": "number",
            "description": "Dry weight of waste in grams recorded during curing",
            "example": 2.5,
            "minimum": 0
          }
        },
        "required": [
          "dryFlowerWeight",
          "dryOtherMaterialWeight",
          "dryWasteWeight"
        ]
      },
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-friendly room name",
            "example": "Flower Room A-1"
          },
          "roomType": {
            "type": "string",
            "description": "Optional room type (e.g., \"flower\", \"veg\", \"mother\", \"dry\")",
            "example": "flower"
          },
          "status": {
            "type": "string",
            "description": "Optional room status (e.g., \"active\", \"inactive\", \"quarantine\")",
            "example": "active"
          },
          "metadata": {
            "type": "object",
            "description": "Optional JSON metadata bag for room (sensors, notes, capacity, etc.)",
            "example": {
              "capacity": 200,
              "climateController": "CC-42",
              "notes": "Near intake"
            },
            "additionalProperties": true
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}